acq/acqsimul/src/test/java/org/dissertation/acqsimul/utils/PanEncryptTest.java
/*
 * MIT License
 *
 * Copyright (c) 2022 Sergio Andres Penen
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 *
 */

package org.dissertation.acqsimul.utils;

import org.dissertation.acqsimul.PanTestUtils;
import org.junit.jupiter.api.Test;

import javax.crypto.*;
import java.security.InvalidAlgorithmParameterException;
import java.security.InvalidKeyException;
import java.security.NoSuchAlgorithmException;
import java.security.spec.InvalidKeySpecException;
import java.util.HexFormat;

import static org.assertj.core.api.Assertions.assertThat;

public class PanEncryptTest {

    @Test
    public void whenEncrypted_thenCanBeDecrypted() throws NoSuchPaddingException, IllegalBlockSizeException, NoSuchAlgorithmException, BadPaddingException, InvalidKeySpecException, InvalidKeyException, InvalidAlgorithmParameterException {
        String encryptedPanBlock = PanEncrypt.process( pan );

        System.out.println( "Encrypted PAN block: " + encryptedPanBlock );

        byte[] iv = HexFormat.of().parseHex(encryptedPanBlock.substring(0,32));
        byte[] encryptedPan = HexFormat.of().parseHex(encryptedPanBlock.substring(32));

        assertThat( PanTestUtils.decryptPan( encryptedPan, iv ) ).isEqualTo( pan );
    }

    char[] pan = "5632890092099001".toCharArray();

}
